+++ websearch /model/model.go
@@ -1,0 +1,4 @@
+package model
+
+// completion
+func Completion() {}
