+++ websearch /result.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS search_results (
+    url TEXT PRIMARY KEY,
+    search_json TEXT NOT NULL,
+    summary TEXT
+);
